move import_cluster test into separate file
